LaTeX reports setup + datasets unignored + current report PDFs
I found a simple LaTeX template on Overleaf for the reports, and I set up an Overleaf project for each task. I set up the language from the assignment instrucitons appropriately for each of the questions, so once you have the current version of the repository all set up on your end it should be as simple as just pasting in the LaTeX for each part of the report's deliverables. I used pdfunite to merge the two together and included them just to verify its all rendering as expected. I noticed that the first task says to clean the datasets, so I unignored the datasets/ folder so that we can make sure we have the same exact versions of the data as we work on the assignment.